pyproject.toml
[project]
name = "nomad-bnb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.15.0",
]
